Get /lists

+++ src/main/java/com/example/games/entities/GameList.java
@@ -10,21 +10,21 @@ public class GameList {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
 
     public GameList(){}
 
-    public GameList(Integer id, String name) {
+    public GameList(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
